sweet alert en borrado

+++ src/public/js/deleteProduct.js
@@ -5,8 +5,26 @@ let id = document.getElementById("productId");
 
 
 deletedProduct.addEventListener('click', (e) => {
-   deleteProductById(id.value);
-   id.value = "";
+    Swal.fire({
+        title: 'Esta seguro?',
+        text: "Esta operación no se puede revertir!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'OK!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Borrado!',
+            'El producto fue eliminado.',
+            'success'
+          )
+          deleteProductById(id.value);
+          id.value = "";
+        }
+      })
+  
 })
 
 const deleteProductById = (item) => {
